ej1 y ej2/app-django/mysite/api_rest/viewsets.py
from rest_framework import viewsets
from . import serializers
from .  import models

class CompanyViewSet(viewsets.ModelViewSet):
    queryset = models.Company.objects.all()
    serializer_class = serializers.CompanySerializer
