ej1 y ej2/app-django/requirements.txt
Django==2.2.20
psycopg2==2.8.3
gunicorn==19.9.0
djangorestframework==3.12.4
Faker==8.0.0
factory-boy==3.2.0